fix : packages, user signout
