+++ src/main/java/com/example/itDa/domain/model/article/ArticleFile.java
@@ -1,6 +1,5 @@
-package com.example.itDa.domain.article;
+package com.example.itDa.domain.model.article;
 
-import com.example.itDa.domain.model.User;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
