+++ src/main/java/com/example/itDa/domain/repository/ArticleFileRepository.java
@@ -1,6 +1,6 @@
 package com.example.itDa.domain.repository;
 
-import com.example.itDa.domain.article.ArticleFile;
+import com.example.itDa.domain.model.article.ArticleFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
