+++ src/main/java/com/example/itDa/dto/request/ArticleRequestDto.java
@@ -1,6 +1,6 @@
-package com.example.itDa.domain.article.request;
+package com.example.itDa.dto.request;
 
-import com.example.itDa.domain.Category;
+import com.example.itDa.domain.model.article.Category;
 import lombok.*;
 
 @Getter
