+++ src/main/java/com/example/itDa/dto/response/ArticleResponseDto.java
@@ -1,7 +1,7 @@
-package com.example.itDa.domain.article.response;
+package com.example.itDa.dto.response;
 
-import com.example.itDa.domain.Category;
-import com.example.itDa.domain.Status;
+import com.example.itDa.domain.model.article.Category;
+import com.example.itDa.domain.model.article.Status;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,9 +13,13 @@ import java.util.List;
 
 @Getter
 @Builder
-@AllArgsConstructor
 @NoArgsConstructor
-public class EditArticleResponseDto {
+@AllArgsConstructor
+public class ArticleResponseDto {
+    private Long userId;
+    private String username;
+    private String profileImg;
+    private int density;
     private Long articleId;
     private String articleName;
     private String substance;
@@ -23,8 +27,13 @@ public class EditArticleResponseDto {
     private Status status;
     private String location;
     @JsonFormat(shape = JsonFormat.Shape.STRING,pattern = "yyyy-MM-dd'T'HH:mm:ss.SS",timezone = "Asia/Seoul")
+    private LocalDateTime createdAt;
+    @JsonFormat(shape = JsonFormat.Shape.STRING,pattern = "yyyy-MM-dd'T'HH:mm:ss.SS",timezone = "Asia/Seoul")
     private LocalDateTime updatedAt;
     private int sellPrice;
     private List<String> fileName;
     private List<String> fileUrl;
+    private boolean like;
+
+
 }
