+++ src/main/java/com/example/itDa/dto/response/EditArticleResponseDto.java
@@ -1,28 +1,21 @@
-package com.example.itDa.domain.article.response;
+package com.example.itDa.dto.response;
 
-import com.example.itDa.domain.Category;
-import com.example.itDa.domain.Status;
-import com.example.itDa.domain.article.Article;
-import com.example.itDa.domain.article.ArticleFile;
+import com.example.itDa.domain.model.article.Category;
+import com.example.itDa.domain.model.article.Status;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Getter
 @Builder
-@NoArgsConstructor
 @AllArgsConstructor
-public class ArticleResponseDto {
-    private Long userId;
-    private String username;
-    private String profileImg;
-    private int density;
+@NoArgsConstructor
+public class EditArticleResponseDto {
     private Long articleId;
     private String articleName;
     private String substance;
@@ -30,13 +23,8 @@ public class ArticleResponseDto {
     private Status status;
     private String location;
     @JsonFormat(shape = JsonFormat.Shape.STRING,pattern = "yyyy-MM-dd'T'HH:mm:ss.SS",timezone = "Asia/Seoul")
-    private LocalDateTime createdAt;
-    @JsonFormat(shape = JsonFormat.Shape.STRING,pattern = "yyyy-MM-dd'T'HH:mm:ss.SS",timezone = "Asia/Seoul")
     private LocalDateTime updatedAt;
     private int sellPrice;
     private List<String> fileName;
     private List<String> fileUrl;
-    private boolean like;
-
-
 }
