+++ src/main/java/com/example/itDa/service/ArticleService.java
@@ -1,17 +1,16 @@
-package com.example.itDa.domain.article.service;
+package com.example.itDa.service;
 
-import com.example.itDa.domain.Category;
-import com.example.itDa.domain.Status;
-import com.example.itDa.domain.article.Article;
-import com.example.itDa.domain.article.ArticleFile;
-import com.example.itDa.domain.article.Like;
-import com.example.itDa.domain.article.repository.ArticleRepository;
-import com.example.itDa.domain.article.repository.LikeRepository;
-import com.example.itDa.domain.article.request.ArticleRequestDto;
-import com.example.itDa.domain.article.request.EditArticleRequestDto;
-import com.example.itDa.domain.article.response.ArticleResponseDto;
-import com.example.itDa.domain.article.response.EditArticleResponseDto;
-import com.example.itDa.domain.article.response.ViewAllArticleResponseDto;
+import com.example.itDa.domain.model.article.Article;
+import com.example.itDa.domain.model.article.ArticleFile;
+import com.example.itDa.domain.model.article.Like;
+import com.example.itDa.domain.model.article.Status;
+import com.example.itDa.domain.repository.ArticleRepository;
+import com.example.itDa.domain.repository.LikeRepository;
+import com.example.itDa.dto.request.ArticleRequestDto;
+import com.example.itDa.dto.request.EditArticleRequestDto;
+import com.example.itDa.dto.response.ArticleResponseDto;
+import com.example.itDa.dto.response.EditArticleResponseDto;
+import com.example.itDa.dto.response.ViewAllArticleResponseDto;
 import com.example.itDa.domain.model.User;
 import com.example.itDa.domain.repository.ArticleFileRepository;
 import com.example.itDa.domain.repository.UserRepository;
